add a timeout for half closed tcp connections in ct
In DSR mode, the first node sometimes forward the pkts to the
real server and set up a ct entry for it. Normally, tcp ct entry
will be deleted when it got two FIN from both ends. But this is not
true for the middle node in DSR. As a result, may ct enties
look like half closed and will last for 6 hours ( default timeout )

This makes the ct table too large and cpu overhead high and may
also cause other issues like pod delete/create failures. To fix
it add a "bpf-ct-timeout-regular-tcp-half-close" option (default
is 5 min) to clean half closed tcp quickly.

Note, still need to take care of backend pods on the 1st node case.
Will do in a later patch

Signed-off-by: Jiang Wang <jiang.wang@bytedance.com>